--- server/db/index.js
@@ -1,15 +1,0 @@
-/* eslint-disable no-console */
-import dotenv from 'dotenv';
-import { Pool } from 'pg';
-
-dotenv.config();
-
-const db = new Pool({
-  connectionString: process.env.DATABASE_URL,
-});
-
-db.on('connect', () => {
-  console.log('database connected');
-});
-
-export default db;
